Add GitInfo package.
This is a package found at:

https://github.com/RWTH-HPC/CMake-gitinfo/blob/master/cmake/FindGitInfo.cmake

which seems to do what's needed for the `CVSTAG` business.

It was about the only package I found that could correctly see that
`@MAPL` was on a *different* tag than `GEOSadas`. I'm sure @tclune could
figure this out, but I was in a bit of a rush, so, for now, I "borrow"
from another expert. The code bits used in GEOSdas_App are (based on
examples from above):

```cmake
if(GIT_FOUND)
   GIT_WC_INFO(${PROJECT_SOURCE_DIR} Project)
endif()

if (NOT "${Project_WC_LATEST_TAG}" STREQUAL "")
   set (GIT_TAG_OR_REV "${Project_WC_LATEST_TAG}")
else ()
   set (GIT_TAG_OR_REV "${Project_WC_REVISION_HASH}")
endif ()

configure_file(CVSTAG.in CVSTAG @ONLY)
install(FILES ${CMAKE_CURRENT_BINARY_DIR}/CVSTAG DESTINATION etc)
```
@tclune might have better suggestions, but it seems to create a `CVSTAG`
with the 7-digit hash (since GEOSadas isn't tagged yet).