Updates for future Python 3 support
In Python 3, the suffix for f2py shared object libraries gets weird.
For example, instead of
```
GFIO_.so
```
you get:
```
GFIO_.cpython-36m-x86_64-linux-gnu.so
```

`UseF2Py.cmake` seems like it should have supported this, but the CMake
was all wonky and a chunk of the code could *never* run. This fixes it.

BUT, to use this code, `F2PY_SUFFIX` should not be set in any
CMakeLists.txt as it will override the defaults.